[MS-162] Feat: 이용내역 API 개발

+++ src/main/java/com/modutaxi/api/domain/roomwaiting/service/RoomWaitingService.java
@@ -39,6 +39,8 @@ public class RoomWaitingService {
     private final MemberRepository memberRepository;
     private final ChatService chatService;
 
+    private static final int FULL_MEMBER = 4;
+
     /**
      * 방 참가 신청
      */
@@ -110,6 +112,13 @@ public class RoomWaitingService {
             throw new BaseException(ParticipateErrorCode.USER_ALREADY_IN_ROOM);
         }
 
+        if(room.getCurrentHeadcount() == FULL_MEMBER){
+            throw new BaseException(ParticipateErrorCode.ROOM_IS_FULL);
+        }
+
+        //fcm구독
+        fcmService.subscribe(participant.getId(), Long.valueOf(roomId));
+
         //대기열에서 제거
         redisChatRoomRepositoryImpl.removeFromWaitingList(roomId, memberId);
         //채팅방에 저장
@@ -120,8 +129,6 @@ public class RoomWaitingService {
         redisChatRoomRepositoryImpl.setUserEnterInfo(memberId, chatRoomMappingInfo);
         room.plusCurrentHeadCount();
 
-        //fcm구독
-        fcmService.subscribe(participant.getId(), Long.valueOf(roomId));
         //참가 수락되었다는 메세지 본인에게 전송
         fcmService.sendPermitParticipate(participant, roomId);
